#2442: Inherited UniverseDefault from Universe interface
